added dry-monades. next step tests
